add dry_run flag

+++ src/cameratransfer/os_output_file_writer.py
@@ -9,6 +9,7 @@ logger = logging.getLogger(__name__)
 class OSOutputFileWriter:
     base_image_location: Path
     base_video_location: Path
+    dry_run: bool = False
 
     def write_file(self, file_name: str, file_last_modified: datetime, file_content: bytes, file_category: str, sub_folder: Path) -> None:
         logger.debug(f"file_name: {file_name}")
@@ -18,6 +19,11 @@ class OSOutputFileWriter:
         elif file_category == "video":
             fq_file_name = self.base_video_location.joinpath(sub_folder, file_name)
         
+        if self.dry_run:
+            logger.debug(f"Would write to file: {fq_file_name}")
+            return
+
+        logger.debug(f"Writing to file: {fq_file_name}")
         fq_file_name.parent.mkdir(parents=True, exist_ok=True)
         fq_file_name.touch(exist_ok=False)
         fq_file_name.write_bytes(file_content)
